[add/#70] 매칭 결과 작성 네비게이션 세팅
