+++ app/src/main/java/com/smashing/app/presentation/write/confirm/ConfirmContract.kt
@@ -1,13 +1,4 @@
 package com.smashing.app.presentation.write.confirm
 
-import androidx.compose.runtime.Immutable
-import com.smashing.app.core.common.type.ReviewRatingType
-import kotlinx.collections.immutable.ImmutableSet
-import kotlinx.collections.immutable.persistentSetOf
-
 sealed interface ConfirmContract {
-    @Immutable
-    data class State(
-        val selectedRatingTypes: ImmutableSet<ReviewRatingType> = persistentSetOf(),
-    )
 }
