--- app/src/main/java/com/smashing/app/presentation/write/navigation/SubmitNavigation.kt
@@ -1,26 +1,0 @@
-package com.smashing.app.presentation.write.navigation
-
-import androidx.navigation.NavController
-import androidx.navigation.NavGraphBuilder
-import androidx.navigation.NavOptions
-import androidx.navigation.compose.composable
-import com.smashing.app.core.common.navigation.Route
-import com.smashing.app.presentation.write.submit.SubmitRoute
-import kotlinx.serialization.Serializable
-
-fun NavController.navigateToSubmit(
-    navOptions: NavOptions? = null
-) = navigate(Submit, navOptions)
-
-fun NavGraphBuilder.submitGraph(
-    navigateUp: () -> Unit,
-) {
-    composable<Submit> {
-        SubmitRoute(
-            navigateUp = navigateUp,
-        )
-    }
-}
-
-@Serializable
-data object Submit : Route
