+++ app/src/main/java/com/smashing/app/presentation/write/submit/SubmitViewModel.kt
@@ -2,8 +2,10 @@ package com.smashing.app.presentation.write.submit
 
 import androidx.compose.foundation.text.input.TextFieldState
 import androidx.lifecycle.ViewModel
+import com.smashing.app.core.common.type.ReviewRatingType
 import com.smashing.app.presentation.write.model.MatchPlayer
 import dagger.hilt.android.lifecycle.HiltViewModel
+import kotlinx.collections.immutable.toImmutableSet
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.flow.update
@@ -17,6 +19,7 @@ class SubmitViewModel @Inject constructor(
 
     val leftTextFieldState: TextFieldState = TextFieldState()
     val rightTextFieldState: TextFieldState = TextFieldState()
+    val reviewTextFieldState: TextFieldState = TextFieldState()
 
     fun updateSelectedWinner(winnerName: String) = _uiState.update { state ->
         val isSubmitterWinner = winnerName == state.submitter.name
@@ -63,6 +66,15 @@ class SubmitViewModel @Inject constructor(
         receiverScore: Int,
     ): Boolean = if (isSubmitterWinner) submitterScore > receiverScore
     else receiverScore > submitterScore
+
+    fun updateSelectedRatingType(type: ReviewRatingType) = _uiState.update { state ->
+        val next = if (type in state.selectedRatingTypes) {
+            state.selectedRatingTypes - type
+        } else {
+            state.selectedRatingTypes + type
+        }
+        state.copy(selectedRatingTypes = next.toImmutableSet())
+    }
 
     private fun getDummyState(): SubmitContract.State {
         return SubmitContract.State(
